Complete mechanism to add card to display on user submission of form
This took a lot of trial and error - to note is you should be careful
of the method with which you call functions from HTML elements. Spent
better part of 1 hour trying to figure out why the box I appended would
disappear after 0.2 seconds. Found out it was because I used <form
onSubmit="return handleClick()">, which for some reason didn't work but
someone recommended this on StackOverflow. I ended up changing this
attribute to action="javascript:handleClick()" and it worked perfectly.